orcarray: Use size_t for pointer arithmetics
Fixes these warnings on windows x86_64:

orcarray.c:80:33: error: cast from pointer to integer of different size [-Werror=pointer-to-int-cast]
orcarray.c:80:22: error: cast to pointer from integer of different size [-Werror=int-to-pointer-cast]

+++ orc-test/orcarray.c
@@ -77,7 +77,7 @@ orc_array_new (int n, int m, int element_size, int misalignment,
 #else
   data = malloc (ar->alloc_len + ALIGNMENT);
   ar->alloc_data = data;
-  ar->aligned_data = (void *)((((unsigned long)data) + (ALIGNMENT-1))&(~(ALIGNMENT-1)));
+  ar->aligned_data = (void *)((((size_t)data) + (ALIGNMENT-1))&(~(ALIGNMENT-1)));
 #endif
 #endif
